thermometer_rpi/requirements.txt
RPi.bme280 == 0.2.*
smbus-cffi == 0.5.*
paho-mqtt == 1.3.*
